Rename `GuardedExistedPathBuf` to `LazyDirectoryCreator` and extract to separate directory
