+++ scarb/src/internal/fsx.rs
@@ -5,10 +5,6 @@ use std::path::{Path, PathBuf};
 
 use anyhow::{anyhow, Context, Result};
 use camino::{Utf8Path, Utf8PathBuf};
-use once_cell::sync::OnceCell;
-use tracing::trace;
-
-use create_output_dir::create_output_dir;
 
 /// Equivalent to [`std::fs::canonicalize`] with better error messages.
 pub fn canonicalize(p: impl AsRef<Path>) -> Result<PathBuf> {
@@ -81,80 +77,5 @@ impl PathBufUtf8Ext for PathBuf {
     fn try_into_utf8(self) -> Result<Utf8PathBuf> {
         Utf8PathBuf::from_path_buf(self)
             .map_err(|path| anyhow!("path `{}` is not UTF-8 encoded", path.display()))
-    }
-}
-
-#[derive(Debug)]
-pub struct GuardedExistedPathBuf<'p> {
-    path: PathBuf,
-    creation_lock: OnceCell<()>,
-    parent: Option<&'p GuardedExistedPathBuf<'p>>,
-    is_output_dir: bool,
-}
-
-impl<'p> GuardedExistedPathBuf<'p> {
-    pub fn new(path: impl Into<PathBuf>) -> Self {
-        Self {
-            path: path.into(),
-            creation_lock: OnceCell::new(),
-            parent: None,
-            is_output_dir: false,
-        }
-    }
-
-    pub fn new_output_dir(path: impl Into<PathBuf>) -> Self {
-        Self {
-            is_output_dir: true,
-            ..Self::new(path)
-        }
-    }
-
-    pub fn child(&'p self, path: impl AsRef<Path>) -> Self {
-        Self {
-            path: self.path.join(path),
-            creation_lock: OnceCell::new(),
-            parent: Some(self),
-            is_output_dir: false,
-        }
-    }
-
-    pub fn as_unchecked(&self) -> &Path {
-        &self.path
-    }
-
-    pub fn into_unchecked(self) -> PathBuf {
-        self.path
-    }
-
-    pub fn as_existent(&self) -> Result<&Path> {
-        self.ensure_created()?;
-        Ok(&self.path)
-    }
-
-    pub fn into_existent(self) -> Result<PathBuf> {
-        self.ensure_created()?;
-        Ok(self.path)
-    }
-
-    fn ensure_created(&self) -> Result<()> {
-        if let Some(parent) = self.parent {
-            parent.ensure_created()?;
-        }
-
-        self.creation_lock
-            .get_or_try_init(|| {
-                trace!(
-                    "creating directory {}; output_dir={}",
-                    &self.path.display(),
-                    self.is_output_dir
-                );
-
-                if self.is_output_dir {
-                    create_output_dir(&self.path)
-                } else {
-                    create_dir_all(&self.path)
-                }
-            })
-            .copied()
     }
 }
