+++ scarb/src/internal/mod.rs
@@ -1,4 +1,5 @@
 pub mod asyncx;
 pub mod fsx;
+pub mod lazy_directory_creator;
 pub mod static_hash_cache;
 pub mod to_version;
